A function to explore the definition of field visits

A function to explore the definition of field visits. Visits are a central concept in the approach to species observation data used by the BIRDS package. In order to assess if your definition of visit aligns with your grid size, you must explore the spatial extent of visits.

exploreVisits(
  x,
  visitCol = NULL,
  sppCol = "scientificName",
  parallel = FALSE,
  nc = NULL
)

Arguments

x
an object of class ‘OrganizedBirds’ (organised BIRDS Spatial data frame). See organizeBirds.

visitCol
name of the column for the visits UID.

sppCol
name of the column for species names.

parallel
logic. Whether to run in parallel (then the package 'parallel' is required). Default is FALSE#'

nc
integer or NULL

Value

a data.frame with summarized data per visit:

“day”

“month”

“year”

“nObs”: number of species observations

“SLL”: species list length (i.e. the number of observed species)

“effortDiam”: the 2 times the maximum of the distances between the centroid of all observation points and any individual observation.

“medianDist”: the median (Q2) of the distances between the centroid and the observations, in meters.

“iqrDist”: the interquartile range of the distances between the centroid the observations, in meters.

“nUniqueLoc”: the number of unique combination of coordinates (locations).

“nClusters”: the number of clusters defined by the DBSCAN algorithm, a minimum of 3 observations per cluster within the median distance between all observations. If the number of clusters is = 0 means that there are at least 3 unique locations but observations are too spread and no cluster was found. If the number of unique locations is less than 3, observations are considered as a single cluster without outliers.

“nOutliers”: the number of observations whose distance to any cluster is beyond the median distance between all observations.

See also

Examples

if(interactive()){
# create a visit-based data object from the original observation-based data
OB <- organizeBirds(bombusObs)
visitStats <- exploreVisits(OB)
# esquisse::esquisser(visitStats)
# alternatively, plot the variable you want, e.g.:
# to see the distribution of distances covered on each visit
hist(visitStats$effortDiam)
# to see the distribution of species list lengths of each visit
hist(visitStats$SLL)
# to identify suspicious visits reported the first day of each month or year
hist(visitStats$day)
# to see correlations
plot(visitStats$nObs, visitStats$effortDiam)
plot(visitStats$SLL, visitStats$effortDiam)
# to see the distributions of observations along the days of the month
plot(visitStats$day, visitStats$nObs)
}