Create ignorance scores

Ignorance scores are a proxy for the lack of sampling effort, computed by making the number of observations relative to a reference number of observations that is considered to be enough to reduce the ignorance score by half (henceforth the Half-ignorance approach). The algorithm behind the Ignorance Score is designed for comparison of bias and gaps in primary biodiversity data across taxonomy, time and space Read more here: Ruete 2015 Biodiv Data J 3:e5361, doi:10.3897/BDJ.3.e5361

exposeIgnorance(nObs, nSpp = NULL, h = 1)

Arguments

nObs
an object of any class (mainly resulting from summariseBirds or exportBirds with the number of observations, or visits in your desired analysis unit.

nSpp
the number of unique species observed

h
the half ignorance parameter value.

Value

a data.frame with ignorance scores

See also

Examples

OB <- organizeBirds(bryophytaObs, sppCol = "scientificName", simplifySppName = TRUE)
grid <- makeGrid(searchPolygon, gridSize = 10)
SB <- summariseBirds(OB, grid=grid)
ignorance <- exposeIgnorance(nObs=SB$spatial@data$nObs)