Make a discrete global grid

Construct a discrete global grid system (dggs) object over a preferred polygon.

makeDggrid(
  polygon,
  gridSize,
  buffer = FALSE,
  topology = "hexagon",
  aperture = 3,
  simplify = FALSE,
  tol = 0.01
)

Arguments

polygon
an object of class ‘SpatialPolygon’ or ‘SpatialPolygonDataFrame’

gridSize
width of the cells in Km. It defines the central assumption of this package that is the maximum area a person can explore during a day. Be aware, that the spatial extent of a visit is dependent on the taxonomic group, and many other variables. Maximum recomended for this package 10 km if there is no reliable definition for the spatial extent for visits.

buffer
shall the grid cells include the polygon border? Then TRUE (default = FALSE).

topology
Shape of cell. shall the grid cells be hexagonal, diamonds or triangular? Options are: “hexagon”, “diamond”, “tirangle”. Default: “hexagon”.

aperture
How finely subsequent resolution levels divide the grid. Options are: 3, 4. Only applicable for topology = "hexagon". Default for topology = "hexagon" is 3, else aperture = 4.

simplify
simplifies the polygon geometry using the Douglas-Peuker algorithm (from rgeos package). Complicated polygons (those with much detail) make this function run slower.

tol
numerical tolerance value for the simplification algorith. Set to 0.01 as default.

Value

an object of class ‘SpatialPolygon’ with a set of polygons conforming to a grid of equal-area cells, with geodesic coordinates in WGS84 (ESPG:4326).

Details

This function depends on a package that is no longer on CRAN. You can find it in its GitHub repository https://github.com/r-barnes/dggridR. Also, this may generate odd results for very large rectangles, because putting rectangles on spheres is weird... as you should know, if you're using this package. Use the function exploreVisits() to assess if your definition of visit aligns with your grid size.

Note

Depending on the total number of grid cells the computations may take time. If there are more than 100 cells on any dimension a warning message will be displayed. Grid cells must be smaller than the sampling area. If the grid cell size is wider than the polygon on any dimension an error message will be displayed.

See also

Examples

# \donttest{
library(dggridR) ## Not anylonger on CRAN

#> Warning: package 'dggridR' was built under R version 3.6.3

#> Loading required package: rgdal

#> rgdal: version: 1.4-8, (SVN revision 845)
#>  Geospatial Data Abstraction Library extensions to R successfully loaded
#>  Loaded GDAL runtime: GDAL 2.2.3, released 2017/11/20
#>  Path to GDAL shared files: 
#>  GDAL binary built with GEOS: TRUE 
#>  Loaded PROJ.4 runtime: Rel. 4.9.3, 15 August 2016, [PJ_VERSION: 493]
#>  Path to PROJ.4 shared files: E:/Alejandro/Documents/R/win-library/3.6/rgdal/proj
#>  Linking to sp version: 1.3-2 

#> Loading required package: ggplot2

#> Loading required package: dplyr

#> Warning: package 'dplyr' was built under R version 3.6.2

#> 
#> Attaching package: 'dplyr'

#> The following object is masked from 'package:testthat':
#> 
#>     matches

#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag

#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union

grid <- makeDggrid(gotaland, gridSize = 10)# }

#> Resolution: 12, Area (km^2): 95.9778454662114, Spacing (km): 9.67579208698331, CLS (km): 11.0545373458563