Makes a grid

Makes a grid adapted to the purpose of this package and simplifing options from the sp package. The central concept of the BIRDS package is the definition of the field visit, and most likely, your grid size will define the maximum area a person can explore during a day. Use the function exploreVisits() to assess if your definition of visit aligns with your grid size.

makeGrid(polygon, gridSize, buffer = FALSE, hexGrid = TRUE,
  offset = NULL, simplify = FALSE, tol = 0.01)

Arguments

polygon
an object of class ‘SpatialPolygon’ or ‘SpatialPolygonDataFrame’

gridSize
width of the cells in Km. It defines the central assumption of this package that is the maximum area a person can explore during a day. Be aware, that the spatial extent of a visit is dependent on the taxonomic group, and many other variables. Maximum recomended for this package 10 km if there is no reliable definition for the spatial extent for visits.

buffer
shall the grid cells include the polygon border? Then TRUE (default = FALSE).

hexGrid
shall the grid cells be hexagonal? Then TRUE (default). Else squared grid cells.

offset
the offset (position) of the grid (from spsample methods). If it is left empty (NULL, default), then takes default values. For squared grid cells the default is set to c(0.5,0.5) ("centric systematic"). For hexagonal grid cells the default is set to c(0,0).

simplify
simplifies the polygon geometry using the Douglas-Peuker algorithm (from rgeos package). Complicated polygons (those with much detail) make this function run slower.

tol
numerical tolerance value for the simplification algorith. Set to 0.01 as default.

Value

an object of class ‘SpatialPolygon’ with a set of polygons conforming to a grid of equal-area cells, with geodesic coordinates in WGS84 (ESPG:4326).

Note

Depending on the total number of grid cells the computations may take time. If there are more than 100 cells on any dimension a warning message will be displayed. Grid cells must be smaller than the sampling area. If the grid cell size is wider than the polygon on any dimension an error message will be displayed.

See also

Examples

grid <- makeGrid(gotaland, gridSize = 10)