Relative observation index (Spatial)

This function extracts the proportion of visits (or observations) detecting a focal species to all visits (or observations) over space.

obsIndexSpatial(x, focalSp = NULL, visits = TRUE, fs.rm = TRUE, norm = TRUE)

Arguments

x
an object of class ‘SummarizeBirds’.

focalSp
the focal sp to look for

visits
if TRUE (default) the observation index is calculated over number of visits, else uses the number of observations.

fs.rm
if TRUE, assumes that the observations for the focal species are included in 'group' and will remove them

norm
if TRUE, the result is nomalized to a 0-1 range

Value

A spatial object