Overlay BIRDS object and grid

Make an overlay for an OrganizedBirds object and a grid to identify which observations that fall into which grid cell.

overlayBirds(x, grid, spillOver = NULL)

# S3 method for OrganizedBirds
overlayBirds(x, grid, spillOver = NULL)

Arguments

x
An OrganizedBirds object

grid
A SpatialPolygonsDataFrame object of the grid over the study area

spillOver
Specifies if the function should search for observations with the same visitUID over several grid cell, and what to do with them. Default is NULL. It also accepts c("unique", "duplicate").

Value

The output is a OverlaidBirds-class object, which is a list containing three objects;

observationsInGrid

Is basically the data in the OrganizedBirds object split by each grid cell (n.b. the use of spillOver = TRUE discussed under "Usage")

grid

The SpatialPolygonsDataFrame from the input, but cleared of data to not waste unnecessary memory

nonEmptyGridCells

An integer vector of which grid cells that have observations

Details

This function takes an OrganizedBirds object, created by organizeBirds, and a polygon grid-layer, which could be created by makeGrid and splits the visits in the OrganizedBirds (i.e. data belonging to identified visits) to each grid cell.

If spillOver = NULL the splitting is done spatially according to the overlay of observations and grid cells, without further consideration of coherence for visits (visit UID). If spillOver = c("unique", "duplicate") the splitting will be done spatially in a first step, and then: if (spillOver = "unique") assigns (and moves) all observations with same visitUID to the grid cell with most observations (or picks one grid cell at random if there is a tie); or if (spillOver = "duplicate") duplicates all observations with same visitUID across all grid cells containing at least one observation with that visitUID.

The later approach is usefull when the amount of observations spilled over neighbouring cells is minimal and information over the spatial extent of the sampling effort is more important than sample independence.

Examples

ob <- organizeBirds(bombusObs)
grid <- makeGrid(gotaland, gridSize = 10)

#> Warning: Discarded ellps WGS 84 in CRS definition: +proj=merc +a=6378137 +b=6378137 +lat_ts=0 +lon_0=0 +x_0=0 +y_0=0 +k=1 +units=m +nadgrids=@null +wktext +no_defs

#> Warning: Discarded datum WGS_1984 in CRS definition

#> Warning: NULL source CRS comment, falling back to PROJ string

#> Warning: +init dropped in PROJ string

#> Warning: CRS object has comment, which is lost in output

ovB <- overlayBirds(ob, grid)

#> 2635 observations did not overlap with the grid and will be discarded.

#> 0.06 % of the visits spill over neighbouring grid cells.