Summarize the OrganizedBirds

Takes a OrganizedBirds-object and a SpatialPolygons*-grid and summarizes it in spatial and temporal dimensions.

summarizeBirds(x, grid, spillOver = NULL)

# S3 method for OrganizedBirds
summarizeBirds(x, grid, spillOver = NULL)

summariseBirds(x, grid, spillOver = NULL)

Arguments

x
An OrgnanizedBirds-object created by organizeBirds

grid
A SpatialPolygons or SpatialPolygonsDataFrame-object whith grid cells for the study area. This variable is optional and can be set to NULL, which will treat the area for all observations as one single grid cell.

spillOver
Specifies if the function should search for observations done during the same visit (identified by visit UID) but that fall outside the grid cell. Default is NULL. It also accepts c("unique", "duplicate"). See Details for more information on how to use this.

Value

A SummarizedBirds-object

Details

If spillOver = NULL the splitting is done spatially according to the overlay of observations and grid cells, without further consideration of coherence for visits (visit UID). If spillOver = c("unique", "duplicate") the splitting will be done spatially in a first step, and then: if (spillOver = "unique") assigns (and moves) all observations with same visitUID to the grid cell with most observations (or picks one grid cell at random if there is a tie); or if (spillOver = "duplicate") duplicates all observations with same visitUID across all grid cells containing at least one observation with that visitUID.

The later approach is usefull when the amount of observations spilled over neighbouring cells is minimal and information over the spatial extent of the sampling effort is more important than sample independence.

Examples

ob<-organizeBirds(bombusObs)
grid <- makeGrid(gotaland, gridSize = 10)
SB <- summarizeBirds(ob, grid)

#> 2635 observations did not overlap with the grid and will be discarded.

#> 0.06 % of the visits spill over neighbouring grid cells.

nObsG<-rowSums(SB$spatioTemporal[,,13,"nObs"], na.rm = FALSE)
nObsG2<-SB$spatial@data$nObs
any(nObsG != nObsG2, na.rm = TRUE) ## Check, two ways to obtain the same

#> [1] FALSE