A function to explore the definition of field visits

A function to explore the definition of field visits. Visits are a central concept in the approach to species observation data used by the BIRDS package. In order to assess if your definition of visit aligns with your grid size, you must explore the spatial extent of visits.

exploreVisits(x, visitCol = attr(x, "visitCol"),
  sppCol = "scientificName")

Arguments

x
an object of class ‘OrganizedBirds’ (organised BIRDS Spatial Dataframe). See organizeBirds.

visitCol
name of the column for the visits UID.

sppCol
name of the column for species names.

Value

a data.frame with summarized data per visit:

“day”

“month”

“year”

“nObs”: number of species observations

“SLL”: species list length (i.e. the number of observed species)

“effortDiam”: the diameter of the minumum circle that covers all points (i.e. locations for species observations), in meters. This is calculated as two times the maximum distance between the centroid of all observation points and any individual observation.

“medianDist”: the median (Q2) of the distances between the centroid and the observations, in meters.

“iqrDist”: the interquartile range of the distances between the centroid the observations, in meters.

“nOutliers”: the number of observations whose distance to the centroid is considered an outlier. Ouliers are defined as distances grater than the Q3 * 1.5 (i.e. length(boxplot.stats(distances)$out) as all distances are positive).

See also

Examples

# \donttest{
# create a visit-based data object from the original observation-based data
OB<-organizeBirds(bombusObs)

#> [1] "just keep going"

visitStats<-exploreVisits(OB)

#> Analysing 6721 visits...Finished analysing 6721 visits.

# open an interactive data explorer
# esquisse::esquisser(visitStats)
# }

# alternativelly, plot the variable you want, e.g.:
# to see the distribution of distances covered on each visit
# hist(visitStat$effortDiam)
# to see the distribution of species list lengths of each visit
# hist(visitStat$SLL)
# to identify suspicious visits reported the first day of each month or year
# hist(visitStat$day)
# to see correlations
# plot(visitStat$nObs, visitStat$effortDiam)
# plot(visitStat$SLL, visitStat$effortDiam)
# to see the ditributions of observations along the days of the month
# plot(visitStat$day, visitStat$nObs)